+++ src/components/banner/index.tsx
@@ -1,6 +1,5 @@
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Autoplay, Pagination, Navigation } from "swiper/modules";
-
 import "swiper/css";
 import "swiper/css/pagination";
 import "swiper/css/navigation";
@@ -8,7 +7,6 @@ import { styled } from "styled-components";
 
 const Banner = () => {
   const bannerSettings = {
-    // spaceBetween: 30,
     modules: [Autoplay, Pagination, Navigation],
     centeredSlides: true,
     autoplay: {
@@ -20,14 +18,10 @@ const Banner = () => {
     <BannerContainer>
       <Swiper {...bannerSettings}>
         <SwiperSlide>
-          <img
-            className="banner"
-            src="/assets/png/banner-euilee.png"
-            alt="배너"
-          />
+          <BannerImg src="/assets/png/banner-euilee.png" alt="배너" />
         </SwiperSlide>
         <SwiperSlide>
-          <img className="banner" src="/assets/png/banner-do0.png" alt="배너" />
+          <BannerImg src="/assets/png/banner-do0.png" alt="배너" />
         </SwiperSlide>
       </Swiper>
     </BannerContainer>
@@ -38,9 +32,10 @@ export default Banner;
 
 const BannerContainer = styled.div`
   padding-left: 20px;
-  .banner {
-    width: 340px;
-    height: 85px;
-    border-radius: 10px;
-  }
+`;
+
+const BannerImg = styled.img`
+  width: 340px;
+  height: 85px;
+  border-radius: 10px;
 `;
