+++ src/types/user.ts
@@ -1,4 +1,18 @@
+import { Timestamp } from "firebase/firestore";
+
 export interface NewUserType {
   nickname: string;
   team: string;
+}
+
+export interface UserType {
+  count: number;
+  date: string;
+  location: string;
+  memo?: string;
+  my?: string;
+  myScore: number;
+  realDate: Timestamp;
+  vs: string;
+  vsScore: number;
 }
